refactor(auth-plugin): extract options/parsing to src/plugins/auth/options.ts (CP-4)
